Finished the left treee display. Only items having books are now displayed

+++ src/main/java/com/popa/books/model/node/BookNode.java
@@ -2,13 +2,13 @@ package com.popa.books.model.node;
 
 public class BookNode extends AbstractNode {
 
-    private int howManyBooks;
+    private long howManyBooks;
 
-    public int getHowManyBooks() {
+    public long getHowManyBooks() {
         return howManyBooks;
     }
 
-    public void setHowManyBooks(int howManyBooks) {
+    public void setHowManyBooks(long howManyBooks) {
         this.howManyBooks = howManyBooks;
     }
 }
