+++ src/main/java/com/popa/books/model/node/EdituraNode.java
@@ -2,22 +2,22 @@ package com.popa.books.model.node;
 
 public class EdituraNode extends AbstractNode {
 
-    private int howManyEdituri;
-    private int howManyBooks;
+    private long howManyEdituri;
+    private long howManyBooks;
 
-    public int getHowManyBooks() {
+    public long getHowManyBooks() {
         return howManyBooks;
     }
 
-    public void setHowManyBooks(int howManyBooks) {
+    public void setHowManyBooks(long howManyBooks) {
         this.howManyBooks = howManyBooks;
     }
 
-    public int getHowManyEdituri() {
+    public long getHowManyEdituri() {
         return howManyEdituri;
     }
 
-    public void setHowManyEdituri(int howManyEdituri) {
+    public void setHowManyEdituri(long howManyEdituri) {
         this.howManyEdituri = howManyEdituri;
     }
 }
